Split Wikidata::Category to a separate file
